requirements.txt
# This file is expected by Heroku.

-r requirements/production.txt

django-debug-toolbar~=3.2.4
pytest~=7.0.1
django-filter~=21.1
